Refactorear testeo. Eliminar archivos innecesarios. Coverage subido a 67%.
