--- src/App/__tests__/App.test.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import Enzyme, { render } from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
-import App from '../App';
-
-Enzyme.configure({ adapter: new Adapter() });
-
-it('renders without crashing', () => {
-  render(<App />);
-});
