+++ src/Seguidor/test/actions.test.js
