+++ src/Seguidor/test/selector.test.js
